requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.0
plotly==5.18.0
schedule==1.2.1
flask
numpy
scikit-learn
transformers
torch
tqdm
watchdog
uvicorn
fastapi
